% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfun.R
\name{myfun}
\alias{myfun}
\title{create an R function with the document. The document is copied from klutometis/roxygen}
\usage{
myfun(x, y)
}
\arguments{
\item{x}{input numeric vector}

\item{y}{input numeric vector}
}
\description{
create an R function with the document. The document is copied from klutometis/roxygen
}
\examples{
str_length(letters)
str_length(c("i", "like", "programming", NA))
}
\seealso{
\code{\link{fun_add}} which this function wraps
}
