% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfun.R
\name{myfun_add}
\alias{myfun_add}
\title{create second R function with the document.}
\usage{
myfun_add(x, y)
}
\arguments{
\item{x}{input numeric vector}

\item{y}{input numeric vector}
}
\value{
sum up the numeric vectors
}
\description{
create second R function with the document.
}
\examples{
str_length(letters)
str_length(c("i", "like", "programming", NA))
}
\seealso{
\code{\link{fun}} which this function wraps
}
